[core] Extract the NetCDF4 format into heat/core/io/netcdf.py
`supports_netcdf`, `load_netcdf`, `save_netcdf` and the
`DNDarray.save_netcdf` binding move out of io.py. The format registers
itself and owns its extension constant and dimension-name template.

`supports_netcdf` is built on `_config.is_installed` and the library is
resolved inside the two functions, so netCDF4 is no longer imported at
`import heat` time. The module-level `__nc_has_par` probe becomes
`_config.netcdf_has_parallel()`, and the "falling back to serial I/O"
warning it gated moves from import time to first use, emitted once on
rank 0.

The write-mode check now goes through `utils.sanitize_write_mode`.

Drops the `"netcdf"` entry from the extension set: it has no leading dot,
and `os.path.splitext` never yields such a value, so it could never match.

`ht.io.nc` -- which the test suite uses -- stays resolvable through the
package's transitional `__getattr__`, now backed by an explicit alias
table rather than by the third-party import happening to be a global of
io.py.
